refactor: Remove legacy search provider
Last usage was in the comment app
