+++ apps/comments/composer/composer/autoload_classmap.php
@@ -23,6 +23,4 @@ return array(
     'OCA\\Comments\\Notification\\Listener' => $baseDir . '/../lib/Notification/Listener.php',
     'OCA\\Comments\\Notification\\Notifier' => $baseDir . '/../lib/Notification/Notifier.php',
     'OCA\\Comments\\Search\\CommentsSearchProvider' => $baseDir . '/../lib/Search/CommentsSearchProvider.php',
-    'OCA\\Comments\\Search\\LegacyProvider' => $baseDir . '/../lib/Search/LegacyProvider.php',
-    'OCA\\Comments\\Search\\Result' => $baseDir . '/../lib/Search/Result.php',
 );
