+++ lib/composer/composer/autoload_classmap.php
@@ -753,9 +753,6 @@ return array(
     'OCP\\Search\\IInAppSearch' => $baseDir . '/lib/public/Search/IInAppSearch.php',
     'OCP\\Search\\IProvider' => $baseDir . '/lib/public/Search/IProvider.php',
     'OCP\\Search\\ISearchQuery' => $baseDir . '/lib/public/Search/ISearchQuery.php',
-    'OCP\\Search\\PagedProvider' => $baseDir . '/lib/public/Search/PagedProvider.php',
-    'OCP\\Search\\Provider' => $baseDir . '/lib/public/Search/Provider.php',
-    'OCP\\Search\\Result' => $baseDir . '/lib/public/Search/Result.php',
     'OCP\\Search\\SearchResult' => $baseDir . '/lib/public/Search/SearchResult.php',
     'OCP\\Search\\SearchResultEntry' => $baseDir . '/lib/public/Search/SearchResultEntry.php',
     'OCP\\Security\\Bruteforce\\IThrottler' => $baseDir . '/lib/public/Security/Bruteforce/IThrottler.php',
